More detailed structure
So the files themselves and where everything is saved is not done yet, but I have moved the respective parts from the PSM into sub-do files.
There's one do file for each matching estimator, a separate one for robustness checks. Feel free to add more do files, as long as you indicate the order in the Master file and name the do-files correctly.